backend/requirements.txt
openai==1.52.0
instructor==1.7.2
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.8.0
pydantic-settings==2.1.0
alembic==1.12.1
psycopg2-binary==2.9.9
aiohttp==3.9.1
certifi==2023.11.17
easyocr==1.7.1
opencv-python-headless==4.8.1.78
numpy==1.26.4
python-dateutil==2.8.2
APScheduler==3.10.4
google-generativeai==0.3.2